[BE] refactor: 게임 서비스 디렉토리 구조 변경 #42
